Add single command conversation as new module
