+++ README.md
@@ -69,22 +69,35 @@ Method          | Parameters     | Returns            | Description
 
 ##Generating the JSON Video File (ffmpeg lib must be installed)
 
-- Option 1: Node.js
+    - Option 1: Node.js - single command
 
-    ```bash
-    cd converter/nodejs
-    node app.js path/to/video/file path/to/video.json/file startTime endTime
-    ```
+        ```bash
+        cd converter/nodejs_one
+        node app.js path/to/video/file path/to/video.json/file startTime endTime
+        ```
 
-- Option 2: PHP
+	- Option 2: Node.js
 
-    - Use ffmpeg to generate the frames from a video file:
+        - Use ffmpeg to generate the frames from a video file:
 
-    ```bash
-    ffmpeg -i video.mp4 -an -f image2 "%d.jpg"
-    ```
+        ```bash
+        ffmpeg -i video.mp4 -an -f image2 "%d.jpg"
+        ```
 
-    - Convert all frames on a single JSON file
+        - Convert all frames on a single JSON file
+
+		```bash
+		cd converter/nodejs
+		node app.js frameStart frameEnd folder/to/imgs/ json/video.json
+		```
+
+	- Option 3: PHP
+
+        - Use ffmpeg to generate the frames from a video file:
+
+        ```bash
+        ffmpeg -i video.mp4 -an -f image2 "%d.jpg"
+        ```
 
 		```bash
 		cd converter/php
