+++ converter/nodejs/package.json
@@ -22,8 +22,6 @@
   },
   "devDependencies": {
     "mime": "^1.2.11",
-    "path": "^0.4.9",
-    "fluent-ffmpeg": "^2.0.1",
-    "mkdirp": "^0.5.1"
+    "path": "^0.4.9"
   }
 }
